requirements.txt
PyYAML>=6.0
numpy>=1.24
requests>=2.31.0
pymilvus>=2.4.0
milvus-lite>=2.4.11
transformers>=4.57.0
torch>=2.1.0
peft>=0.10.0
accelerate>=0.30.0
qwen-vl-utils>=0.0.14
gradio>=4.44.0
rank-bm25>=0.2.2
pypdf>=4.2.0
Pillow>=10.0.0
modelscope>=1.34.0
vllm>=0.11.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
streamlit>=1.39.0
